MTP: Fix leaked weak reference on monitors.
We use a weak reference to automatically clean up the monitor
hash table when monitors are finalized. However, we still need
to remember to remove the weak references when the backend is
itself finalized - otherwise the reference might be fired later
if a monitor happens to get finalized before the backend exits
completely.